+++ src/app/playground/state-management90/_components/OptionWrapper.tsx
@@ -36,7 +36,7 @@ const OptionWrapper: React.FC<Props> = memo(
           transform: CSS.Transform.toString(sortable.transform),
           transition: sortable.transition,
         }}
-        className={twMerge("m-1 border p-1", isDragging && "bg-blue-50")}
+        className={twMerge("my-1 border p-1", isDragging && "bg-blue-50")}
       >
         {/* ドラッグアンドドロップのグリップ */}
         <div className="flex space-x-2">
@@ -55,7 +55,6 @@ const OptionWrapper: React.FC<Props> = memo(
             option={option}
             getOptimisticLatestData={getOptimisticLatestData}
             onUpdateDefaultOption={onUpdateDefaultOption}
-            isDragging={isDragging}
           />
         </div>
       </div>
