+++ src/app/playground/state-management90/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useRef, useCallback, useMemo } from "react";
-import QuestionWrapper from "./_components/QuestionWrapper";
+import PageContent from "./_components/PageContent";
 import { produce, Draft } from "immer";
 import useAuthenticatedGetRequest from "@/app/_hooks/useAuthenticatedGetRequest";
 import dev from "@/app/_utils/devConsole";
@@ -25,6 +25,7 @@ import { StatusCodes } from "@/app/_utils/extendedStatusCodes";
 import useAuth from "@/app/_hooks/useAuth";
 import useConfirmDialog from "@/app/_hooks/useConfirmDialog";
 import { ConfirmDialog } from "@/app/_components/elements/ConfirmDialog";
+import ActionButton from "@/app/_components/elements/ActionButton";
 
 const Page: React.FC = () => {
   const id = "cm1dmmv0s0002dg0zwgbt5vna"; // TODO: デバッグ用
@@ -34,6 +35,8 @@ const Page: React.FC = () => {
   const { apiRequestHeader } = useAuth();
   const dataRef = useRef<SessionEditableFields>();
   const confirmDeleteDialog = useConfirmDialog();
+
+  const [isAddingQuestion, setIsAddingQuestion] = React.useState(false);
 
   // prettier-ignore
   const postApiCaller = useMemo(() => createPostRequest<AddQuestionRequest, ApiResponse<QuestionEditableFields>>(),[]);
@@ -47,11 +50,12 @@ const Page: React.FC = () => {
   //【設問の追加】
   const addQuestion = useCallback(async () => {
     // [POST] /api/v1/teacher/questions/new
+    setIsAddingQuestion(true);
     const ep = `/api/v1/teacher/questions/new`;
     const reqBody: AddQuestionRequest = addQuestionRequestSchema.parse({
       sessionId: data?.data?.id!,
-      order: 1, // TODO: 仮
-      title: `設問0${dataRef.current!.questions!.length! + 1}`, // TODO: 仮
+      order: dataRef.current?.questions.length! + 1,
+      title: `設問0${dataRef.current?.questions.length! + 1}`,
     });
     dev.console.log("■ >>> " + JSON.stringify(reqBody, null, 2));
     const res = await postApiCaller(ep, reqBody, apiRequestHeader);
@@ -70,6 +74,7 @@ const Page: React.FC = () => {
         .build(),
       false
     );
+    setIsAddingQuestion(false);
   }, [apiRequestHeader, data?.data?.id, mutate, postApiCaller]);
 
   //【設問の削除（本体）】
@@ -122,13 +127,6 @@ const Page: React.FC = () => {
       <div className="mb-4 flex space-x-2">
         <button
           className="rounded-md border px-3 py-1 text-sm"
-          onClick={addQuestion}
-        >
-          設問追加
-        </button>
-
-        <button
-          className="rounded-md border px-3 py-1 text-sm"
           onClick={revalidate}
         >
           再取得（再検証）
@@ -136,22 +134,30 @@ const Page: React.FC = () => {
       </div>
 
       {/* 設問 */}
-      <div>
-        {dataRef.current.questions.map((question) => (
-          <QuestionWrapper
-            key={question.id}
-            question={question}
-            getOptimisticLatestData={getOptimisticLatestData}
-            confirmDeleteQuestion={confirmDeleteQuestion}
-          />
-        ))}
+      <div className="mb-4">
+        <PageContent
+          session={dataRef.current}
+          getOptimisticLatestData={getOptimisticLatestData}
+          confirmDeleteQuestion={confirmDeleteQuestion}
+        />
+      </div>
+
+      <div className="mb-4 flex justify-end space-x-2">
+        <ActionButton
+          type="button"
+          variant="add"
+          isBusy={isAddingQuestion}
+          onClick={addQuestion}
+        >
+          設問追加
+        </ActionButton>
       </div>
 
       <ConfirmDialog {...confirmDeleteDialog} />
 
-      <pre className="mt-10 text-xs">
+      {/* <pre className="mt-10 text-xs">
         {JSON.stringify(dataRef.current, null, 2)}
-      </pre>
+      </pre> */}
     </div>
   );
 };
